requirements.txt
Faker==4.0.1
kafka-python==2.0.1
